Replace process events from SmartHome to EventManager
